feat(#13): Adjusted form and Api configuration
