+++ package.json
@@ -13,7 +13,6 @@
     "@types/node": "20.4.1",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
-    "axios": "^1.5.0",
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.9",
     "i18next": "^23.2.11",
